Streamlining belt corner placement
Belt corners are now updated during a drag rather than at the end of the drag. And the drag system (cell selection) now maintains a record of the belt type to be placed during the drag, rather than being computed at the end. This should now be more easily extensible for adding more belt (node) types such as intersections.